# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AirDnD API
  version: 0.1.0
  description: AirDnD 프론트엔드와 백엔드가 공유하는 API 계약입니다.
servers:
  - url: http://127.0.0.1:8080
    description: 로컬 백엔드 서버
paths:
  /api/auth/me:
    get:
      tags: [Auth]
      summary: 현재 로그인 사용자 조회
      responses:
        "200":
          description: 로그인 사용자 또는 null
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/User"
                  - type: object
                    nullable: true
                    description: 비로그인 상태에서는 JSON null을 반환합니다.
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: 로그아웃
      security:
        - cookieAuth: []
      responses:
        "204":
          description: 로그아웃 완료
  /oauth2/authorization/{provider}:
    get:
      tags: [Auth]
      summary: OAuth 로그인 시작
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google]
      responses:
        "302":
          description: OAuth provider로 리다이렉트
  /api/rooms:
    get:
      tags: [Rooms]
      summary: 숙소 목록 검색
      description: >
        숙소 목록을 검색합니다. 텍스트/필터 조건(region, 날짜, 인원, 가격, 반려동물)에
        더해, 지도 화면에서는 현재 보이는 영역(뷰포트)의 경계 좌표를 함께 전달해
        해당 사각형 영역 안의 숙소만 받아올 수 있습니다.


        지도 연동 동작:
          - `south`/`west`/`north`/`east` 4개 경계 좌표는 **모두 함께** 전달해야 합니다.
            하나라도 빠지면 경계 필터는 무시됩니다.
          - 4개가 모두 있으면 `south <= latitude <= north` 이고
            `west <= longitude <= east` 인 숙소만 반환합니다.
          - 응답 개수는 `limit`(기본 200)으로 제한됩니다. 초과 시 잘리므로
            클라이언트는 더 좁은 영역으로 다시 검색하도록 안내해야 합니다.
          - 지도 마커를 찍으려면 응답의 각 숙소에 `latitude`/`longitude`가
            포함되어야 합니다(아래 RoomSummary 참고).
      parameters:
        - name: region
          in: query
          schema:
            type: string
        - name: checkIn
          in: query
          schema:
            type: string
            format: date
        - name: checkOut
          in: query
          schema:
            type: string
            format: date
        - name: guests
          in: query
          schema:
            type: integer
            minimum: 1
        - name: adults
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 8
        - name: children
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 8
        - name: infants
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 8
        - name: minPrice
          in: query
          schema:
            type: integer
            minimum: 0
        - name: maxPrice
          in: query
          schema:
            type: integer
            minimum: 0
        - name: allowsPets
          in: query
          schema:
            type: boolean
        - name: south
          in: query
          description: >
            지도 뷰포트 남쪽(최소) 위도. south/west/north/east 4개를 함께 보내면
            그 사각형 영역 안의 숙소만 반환합니다. 프론트엔드 `MapBounds.south`와 1:1.
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: west
          in: query
          description: 지도 뷰포트 서쪽(최소) 경도. `MapBounds.west`와 1:1.
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: north
          in: query
          description: 지도 뷰포트 북쪽(최대) 위도. `MapBounds.north`와 1:1.
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: east
          in: query
          description: 지도 뷰포트 동쪽(최대) 경도. `MapBounds.east`와 1:1.
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: limit
          in: query
          description: >
            한 번에 반환할 최대 숙소 수. 지도 마커 폭주와 페이로드 비대를 막기 위한
            상한이며, 초과분은 잘립니다. 미지정 시 200.
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 200
      responses:
        "200":
          description: 숙소 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoomSummary"
  /api/rooms/{roomId}:
    get:
      tags: [Rooms]
      summary: 숙소 상세 조회
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: 숙소 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/reservations:
    get:
      tags: [Reservations]
      summary: 내 예약 목록 조회
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 예약 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
        "401":
          $ref: "#/components/responses/Unauthenticated"
    post:
      tags: [Reservations]
      summary: 예약 생성
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationRequest"
      responses:
        "201":
          description: 생성된 예약(PENDING 홀드)의 id
          content:
            application/json:
              schema:
                type: integer
                format: int64
                description: 생성된 예약 id. 이후 /checkout/{reservationId} 결제 흐름으로 이어진다.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/reservations/{reservationId}:
    get:
      tags: [Reservations]
      summary: 예약 상세 조회
      security:
        - cookieAuth: []
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 예약 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Reservations]
      summary: 예약 취소
      security:
        - cookieAuth: []
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 취소 완료
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/payments/orders:
    post:
      tags: [Payments]
      summary: PayPal 주문 생성
      description: >
        PENDING 예약(reservationId)에 대해 PayPal Orders v2 주문을 만든다.
        결제 금액은 클라이언트가 보내지 않고 서버가 해당 예약에서 재계산한다.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentOrderRequest"
      responses:
        "200":
          description: 생성된 PayPal 주문
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrderResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/payments/orders/{orderId}/capture:
    post:
      tags: [Payments]
      summary: 결제 확정(capture)
      description: PayPal 주문을 capture 하고 대상 예약을 PENDING → CONFIRMED 로 확정한다.
      security:
        - cookieAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          description: PayPal Orders v2 order id
          schema:
            type: string
      responses:
        "200":
          description: 확정된 예약 id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/host/rooms:
    get:
      tags: [Host]
      summary: 호스트 숙소 목록 조회
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 호스트 숙소 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostRoom"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Host]
      summary: 호스트 숙소 등록
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostRoomRequest"
      responses:
        "201":
          description: 등록된 숙소
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostRoom"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/host/rooms/{roomId}:
    get:
      tags: [Host]
      summary: 호스트 숙소 상세 조회
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: 호스트 숙소 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostRoom"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Host]
      summary: 호스트 숙소 수정
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostRoomRequest"
      responses:
        "200":
          description: 수정된 숙소
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostRoom"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/host/rooms/{roomId}/status:
    patch:
      tags: [Host]
      summary: 호스트 숙소 상태 변경
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [ACTIVE, INACTIVE, PENDING_APPROVAL]
      responses:
        "200":
          description: 상태 변경된 숙소
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostRoom"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/admin/dashboard:
    get:
      tags: [Admin]
      summary: 관리자 대시보드 지표 조회
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 관리자 대시보드 지표
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboard"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/rooms/pending:
    get:
      tags: [Admin]
      summary: 승인 대기 숙소 목록
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 승인 대기 숙소 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostRoom"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/rooms/{roomId}/approve:
    post:
      tags: [Admin]
      summary: 숙소 승인
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: 승인된 숙소
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostRoom"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/rooms/{roomId}/reject:
    post:
      tags: [Admin]
      summary: 숙소 반려
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: 반려된 숙소
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostRoom"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/users:
    get:
      tags: [Admin]
      summary: 사용자 관리 목록
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 사용자 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/reservations:
    get:
      tags: [Admin]
      summary: 관리자 예약 현황
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 예약 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/waitlist:
    get:
      tags: [Admin]
      summary: 대기 시스템 상태
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 대기열 스냅샷
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistSnapshot"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/notifications:
    get:
      tags: [Notifications]
      summary: 알림 목록
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 알림 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          $ref: "#/components/responses/Unauthenticated"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: integer
  responses:
    BadRequest:
      description: 요청 형식 오류
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthenticated:
      description: 로그인 필요
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 권한 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: 비즈니스 충돌
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [GUEST, HOST, ADMIN]
        avatarUrl:
          type: string
          format: uri
    RoomSummary:
      type: object
      required:
        [id, name, region, address, pricePerNight, rating, reviewCount, maxGuests, imageUrl, isAvailable, allowsPets]
      properties:
        id:
          type: integer
        name:
          type: string
        region:
          type: string
        address:
          type: string
        pricePerNight:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        maxGuests:
          type: integer
        imageUrl:
          type: string
          format: uri
        isAvailable:
          type: boolean
        allowsPets:
          type: boolean
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          description: >
            지도 마커 표시용 위도. 지도 검색을 지원하려면 목록 응답에도 포함해야 합니다.
            (백엔드 RoomResponse가 추가하기 전까지는 누락될 수 있어 required 아님)
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          description: 지도 마커 표시용 경도. latitude와 함께 제공되어야 합니다.
    RoomDetail:
      allOf:
        - $ref: "#/components/schemas/RoomSummary"
        - type: object
          required: [description, amenities, hostName, latitude, longitude]
          properties:
            description:
              type: string
            amenities:
              type: array
              items:
                type: string
            imageUrls:
              type: array
              items:
                type: string
                format: uri
            hostName:
              type: string
            latitude:
              type: number
            longitude:
              type: number
    Reservation:
      type: object
      required: [id, roomId, roomName, roomUrl, checkIn, checkOut, guests, pricePerNight, totalPrice, status]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        roomName:
          type: string
        roomUrl:
          type: string
          format: uri
          description: 숙소 대표 이미지 URL
        region:
          type: string
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        guests:
          type: integer
        pricePerNight:
          type: integer
        totalPrice:
          type: integer
          format: int64
          description: KRW 총액. 서버가 pricePerNight × 박수로 계산하며 클라이언트 입력을 신뢰하지 않는다.
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED]
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: >
            PENDING 홀드 만료 시각. 이 시각 이후의 PENDING 은 점유로 보지 않는다(결제 미완료 자동 해제).
            CONFIRMED/CANCELLED 예약에서는 null.
        guestName:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateReservationRequest:
      type: object
      required: [roomId, checkInDate, checkOutDate, adultCount]
      description: >
        예약(PENDING 홀드) 생성 요청. totalPrice 는 받지 않는다 —
        서버가 숙소 가격과 박수로 계산한다(가격 위변조 방지). guestId 는 인증 세션에서 가져온다.
      properties:
        roomId:
          type: integer
        checkInDate:
          type: string
          format: date
        checkOutDate:
          type: string
          format: date
        adultCount:
          type: integer
          minimum: 1
        childCount:
          type: integer
          minimum: 0
          default: 0
        infantCount:
          type: integer
          minimum: 0
          default: 0
        hasPets:
          type: boolean
          default: false
    CreatePaymentOrderRequest:
      type: object
      required: [reservationId]
      properties:
        reservationId:
          type: integer
          format: int64
          description: 결제 대상 PENDING 예약 id
    CreateOrderResponse:
      type: object
      required: [orderId]
      properties:
        orderId:
          type: string
          description: PayPal Orders v2 order id (capture 시 사용)
    CaptureResponse:
      type: object
      required: [reservationId]
      properties:
        reservationId:
          type: integer
          format: int64
          description: 결제로 확정(CONFIRMED)된 예약 id
    HostRoom:
      allOf:
        - $ref: "#/components/schemas/RoomDetail"
        - type: object
          required: [countryCode, status]
          properties:
            countryCode:
              type: string
              pattern: "^[A-Z]{2}$"
            status:
              type: string
              enum: [ACTIVE, INACTIVE, PENDING_APPROVAL]
    HostRoomRequest:
      type: object
      required:
        [name, region, address, pricePerNight, maxGuests, imageUrl, allowsInfants, allowsPets, countryCode, latitude, longitude]
      properties:
        name:
          type: string
        region:
          type: string
        address:
          type: string
        description:
          type: string
        pricePerNight:
          type: integer
          minimum: 1
        maxGuests:
          type: integer
          minimum: 1
        imageUrl:
          type: string
          format: uri
        imageUrls:
          type: array
          items:
            type: string
            format: uri
        allowsPets:
          type: boolean
          default: false
        allowsInfants:
          type: boolean
          default: false
        countryCode:
          type: string
          pattern: "^[A-Z]{2}$"
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        amenities:
          type: array
          items:
            type: string
    AdminDashboard:
      type: object
      required: [pendingRooms, activeUsers, reservationsToday, waitQueueSize]
      properties:
        pendingRooms:
          type: integer
        activeUsers:
          type: integer
        reservationsToday:
          type: integer
        waitQueueSize:
          type: integer
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        detail:
          type: object
          additionalProperties: true
    AdminUser:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [reservationCount, joinedAt, status]
          properties:
            reservationCount:
              type: integer
            joinedAt:
              type: string
              format: date
            status:
              type: string
              enum: [ACTIVE, BLOCKED]
    WaitlistSnapshot:
      type: object
      required: [status, waitingUsers, averageWaitMinutes, admissionRatePerMinute, updatedAt]
      properties:
        status:
          type: string
          enum: [OPEN, PAUSED, CLOSED]
        waitingUsers:
          type: integer
        averageWaitMinutes:
          type: integer
        admissionRatePerMinute:
          type: integer
        updatedAt:
          type: string
          format: date-time
    Notification:
      type: object
      required: [id, type, title, message, read, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [RESERVATION, HOST, SYSTEM]
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time